requirements.txt
agate==1.6.3
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
dbt-bigquery==1.5.3
dbt-core==1.5.2
dbt-extractor==0.4.1
Flask==2.3.2
future==0.18.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.3
google-cloud-dataproc==5.4.2
google-cloud-logging==3.6.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==21.2.0
hologram==0.0.16
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.1.3
mashumaro==3.6
minimal-snowplow-tracker==0.0.2
msgpack==1.0.5
networkx==2.8.8
packaging==23.1
parsedatetime==2.4
pathspec==0.11.1
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
python-dateutil==2.8.2
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.10
rsa==4.9
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
